Refactor bookmark plugins
Mainly moving functions and methods to utils
