+++ src/plugins/bookmark-views/index.js
@@ -1,5 +1,4 @@
 /**
- * @module converse-bookmark-views
  * @description Converse.js plugin which adds views for XEP-0048 bookmarks
  * @copyright 2020, the Converse.js contributors
  * @license Mozilla Public License (MPLv2)
@@ -7,10 +6,9 @@
 import '@converse/headless/plugins/muc/index.js';
 import BookmarkForm from './form.js';
 import BookmarksView from './bookmarks-list.js';
-import { __ } from 'i18n';
 import { _converse, api, converse } from '@converse/headless/core';
-import { bookmarkableChatRoomView, eventMethods } from './mixins.js';
-import { checkBookmarksSupport } from '@converse/headless/plugins/bookmarks/utils';
+import { bookmarkableChatRoomView } from './mixins.js';
+import { getHeadingButtons, removeBookmarkViaEvent, addBookmarkViaEvent } from './utils.js';
 
 import './styles/bookmarks.scss';
 
@@ -23,8 +21,6 @@ converse.plugins.add('converse-bookmark-views', {
      * If the setting "strict_plugin_dependencies" is set to true,
      * an error will be raised if the plugin is not found. By default it's
      * false, which means these plugins are only loaded opportunistically.
-     *
-     * NB: These plugins need to have already been loaded via require.js.
      */
     dependencies: ['converse-chatboxes', 'converse-muc', 'converse-muc-views'],
 
@@ -37,34 +33,15 @@ converse.plugins.add('converse-bookmark-views', {
             hide_open_bookmarks: true
         });
 
-        Object.assign(_converse, eventMethods);
+        _converse.removeBookmarkViaEvent = removeBookmarkViaEvent;
+        _converse.addBookmarkViaEvent = addBookmarkViaEvent;
+
         Object.assign(_converse.ChatRoomView.prototype, bookmarkableChatRoomView);
 
         _converse.MUCBookmarkForm = BookmarkForm;
         _converse.BookmarksView = BookmarksView;
 
-        /************************ BEGIN Event Handlers ************************/
-        api.listen.on('getHeadingButtons', (view, buttons) => {
-            if (_converse.allow_bookmarks && view.model.get('type') === _converse.CHATROOMS_TYPE) {
-                const bookmarked = view.model.get('bookmarked');
-                const data = {
-                    'i18n_title': bookmarked ? __('Unbookmark this groupchat') : __('Bookmark this groupchat'),
-                    'i18n_text': bookmarked ? __('Unbookmark') : __('Bookmark'),
-                    'handler': ev => view.toggleBookmark(ev),
-                    'a_class': 'toggle-bookmark',
-                    'icon_class': 'fa-bookmark',
-                    'name': 'bookmark'
-                };
-                const names = buttons.map(t => t.name);
-                const idx = names.indexOf('details');
-                const data_promise = checkBookmarksSupport().then(s => (s ? data : ''));
-                return idx > -1
-                    ? [...buttons.slice(0, idx), data_promise, ...buttons.slice(idx)]
-                    : [data_promise, ...buttons];
-            }
-            return buttons;
-        });
-
+        api.listen.on('getHeadingButtons', getHeadingButtons);
         api.listen.on('chatRoomViewInitialized', view => view.setBookmarkState());
     }
 });
